+++ src/contracts/SnappyPdf.php
@@ -13,6 +13,8 @@ namespace enupal\snapshot\contracts;
 use Craft;
 use craft\web\Response;
 use enupal\snapshot\Snapshot;
+use enupal\stripe\elements\Order;
+use enupal\stripe\Stripe;
 use Knp\Snappy\GeneratorInterface;
 use Knp\Snappy\Pdf;
 
@@ -86,6 +88,24 @@ class SnappyPdf extends BaseSnappy
         $html = Craft::$app->getView()->renderTemplate($template, $variables);
 
         return $this->displayHtml($html, $settings);
+    }
+
+    /**
+     * @param Order $order
+     * @param array $settings display inline | url
+     *
+     * @return string|Response
+     * @throws \Throwable
+     * @throws \yii\base\Exception
+     * @throws \yii\web\ServerErrorHttpException
+     */
+    public function displayOrder(Order $order, $settings = null)
+    {
+        $settings['variables']['order'] = $order;
+        $template = $this->getStripePaymentsOrderTemplate($settings);
+        $settings['filename'] = $this->getStripePaymentsFilename($settings);
+
+        return $this->displayTemplate($template, $settings);
     }
 
     /**